Dataset Source file
This dataset is a list of all the medal winners in the Summer Olympics from 1976 Montreal to 2008 Beijing. It includes each and every medal awarded within the period. This dataset is intended for beginners so that they can get a taste of advanced Excel functions which is perhaps one of the key skills required to be a great data scientist. I too got my hands dirty with the dataset and played with some advanced Excel functions. Further, this dataset can also be used for a predictive model as to which country is likely to fetch the highest number of gold in a particular sports category (just an example), etc.

Source file Structure

 ● City: The city where the Olympics took place.
 ● Year: The year of the Olympics.
 ● Sport: The sport the event is categorized under.
 ● Discipline: A subcategory of the sport.
 ● Event: The specific event within a discipline.
 ● Athlete: The name of the athlete who participated.
 ● Gender: The gender of the athlete.
 ● Country_Code: The country code (abbreviation).
 ● Country: The full name of the country.
 ● Event_gender: The gender category of the event.
 ● Medal: The medal won (Gold, Silver, Bronze).